pyproject.toml
[project]
name = "ics"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = "^3.12"
dependencies = ["fastapi (>=0.115.12,<0.116.0)", "gunicorn (>=23.0.0,<24.0.0)", "pydantic[email] (>=2.11.4,<3.0.0)", "pydantic-settings (>=2.9.1,<3.0.0)", "pytest (>=8.3.5,<9.0.0)", "pytest-asyncio (>=0.26.0,<0.27.0)", "tortoise-orm[asyncmy] (>=0.25.0,<0.26.0)", "aerich (>=0.8.2,<0.9.0)", "cryptography (>=44.0.3,<45.0.0)", "python-dotenv (>=1.1.0,<2.0.0)", "tomlkit (>=0.13.2,<0.14.0)", "uvicorn (>=0.34.2,<0.35.0)", "bcrypt (==4.0.1)", "python-jose (>=3.4.0,<4.0.0)", "fastapi-mail (>=1.4.2,<2.0.0)", "status (>=0.2.5,<0.3.0)", "python-multipart (>=0.0.20,<0.0.21)", "types-python-jose (>=3.4.0.20250224,<4.0.0.0)", "types-passlib (>=1.7.7.20250408,<2.0.0.0)"]

[[project.authors]]
name = "jiwon"
email = "realjiwon98@naver.com"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
coverage = "^7.8.0"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
python_version = 3.12
strict = true
plugins = ["pydantic.mypy"]

[tool.aerich]
tortoise_orm = "app.configs.database_settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
